requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
exceptiongroup==1.3.1
fastapi==0.115.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
psycopg==3.2.1
psycopg-binary==3.2.1
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.6
tomli==2.4.0
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
